pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "apicultura"
version = "0.1.0"
description = ""
authors = ["Armando Solheiro <avgsolheiro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.111.0"
pydantic = {extras = ["email"], version = "^2.7.4"}
sqlalchemy = "^2.0.31"
isort = "^5.13.2"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
ruff = "^0.4.10"
httpx = "^0.27.0"
ignr = "^2.2" 
factory-boy = "^3.3.0"
freezegun = "^1.5.1"

[tool.poetry.scripts]
pyright = "pyright"

[tool.ruff]
line-length = 79
extend-exclude =  [
    "migrations",
    "__init__.py",
    "env.py",
    ]

[tool.ruff.format]
preview = true
quote-style = 'double'

[tool.coverage.run]
omit = [
    "apicultura/tests/"
]
